create file for constants
