+++ app/components/listing/feedback/FeedbackSteps.jsx
@@ -88,30 +88,3 @@ export const NavigationButtons = ({
     )}
   </div>
 );
-
-export const TAG_LIST = [
-  {
-    tag: 'Contact Information',
-    selected: false,
-  },
-  {
-    tag: 'Hours',
-    selected: false,
-  },
-  {
-    tag: 'Address',
-    selected: false,
-  },
-  {
-    tag: 'Website Link',
-    selected: false,
-  },
-  {
-    tag: 'Information missing',
-    selected: false,
-  },
-  {
-    tag: 'Other',
-    selected: false,
-  },
-].map(tag => Object.freeze(tag));
