+++ app/components/listing/feedback/constants.js
@@ -1,0 +1,29 @@
+export const TAG_LIST = [
+  {
+    tag: 'Contact Information',
+    selected: false,
+  },
+  {
+    tag: 'Hours',
+    selected: false,
+  },
+  {
+    tag: 'Address',
+    selected: false,
+  },
+  {
+    tag: 'Website Link',
+    selected: false,
+  },
+  {
+    tag: 'Information missing',
+    selected: false,
+  },
+  {
+    tag: 'Other',
+    selected: false,
+  },
+].map(tag => Object.freeze(tag));
+
+export const UPVOTE = 'upvote';
+export const DOWNVOTE = 'downvote';
